Added Drowned Item stacks, will work on drowned listeners later.
