+++ src/main/java/xyz/roosterseatyou/mobitems/itemstacks/undead/zombie/ZombieFeet.java
@@ -1,4 +1,4 @@
-package xyz.roosterseatyou.mobitems.itemstacks.undead.drowned;
+package xyz.roosterseatyou.mobitems.itemstacks.undead.zombie;
 
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.TextColor;
@@ -12,29 +12,28 @@ import org.bukkit.inventory.meta.LeatherArmorMeta;
 import java.util.ArrayList;
 import java.util.List;
 
-public class DrownedMask {
-    public static Component NAME = Component.text("Drowned Mask").color(TextColor.fromHexString("#89E2C7"));
-    public static ItemStack DROWNED_MASK;
+public class ZombieFeet{
+    public static Component NAME = Component.text("Zombie Feet").color(TextColor.fromHexString("#05540d"));
+    public static ItemStack ZOMBIE_FEET;
 
     public static ItemStack init(){
-        drownedMask();
+        zombieFeet();
         return null;
     }
 
-    public static void drownedMask(){
-        ItemStack item = new ItemStack(Material.LEATHER_HELMET);
+    public static void zombieFeet(){
+        ItemStack item = new ItemStack(Material.LEATHER_BOOTS);
         LeatherArmorMeta meta = (LeatherArmorMeta) item.getItemMeta();
         List<Component> lore = new ArrayList<>();
         meta.displayName(NAME);
-        meta.setColor(Color.fromRGB(54, 89, 78));
+        meta.setColor(Color.GREEN);
         meta.addEnchant(Enchantment.PROTECTION_ENVIRONMENTAL, 5, true);
-        lore.add(Component.text("Ironically, you now know how to swim..."));
-        //see UndeadArmorUtils & UnderwaterArmorUtils
-        lore.add(Component.text("ENTITY ID: Drowned").color(TextColor.fromHexString("#89E2C7")).decorate(TextDecoration.ITALIC));
+        lore.add(Component.text("Made from the flesh of Zombies..."));
+        //see UndeadArmorUtils
+        lore.add(Component.text("ENTITY ID: ZOMBIE").color(TextColor.fromHexString("#2d3233")).decorate(TextDecoration.ITALIC));
         lore.add(Component.text("CLASS ID: UNDEAD").color(TextColor.fromHexString("#2d3233")).decorate(TextDecoration.ITALIC));
-        lore.add(Component.text("CLASS ID: UNDERWATER").color(TextColor.fromHexString("#92D2E2")).decorate(TextDecoration.ITALIC));
         meta.lore(lore);
         item.setItemMeta(meta);
-        DROWNED_MASK = item;
+        ZOMBIE_FEET = item;
     }
 }
